\name{bowie}
\alias{bowie}
\docType{data}
\title{
  Topo: Bowie Seamount Marine Protected Area
}
\description{
  A PolySet describing the area of the Pacific Ocean -- consisting of the seabed, the subsoil and 
  the water column above the seabed -- that is designated as the Bowie Seamount Marine Protected Area.
}
\usage{data(bowie)}
\format{
  A data frame with 8 observations on the following 4 variables.
  \describe{
    \item{\code{PID}}{primary polygon ID}
    \item{\code{POS}}{vertex position number}
    \item{\code{X}}{longitude coordinate (degrees W)}
    \item{\code{Y}}{latitude coordinate (degrees N)}
  }
}
\details{
  The area of the Pacific Ocean -- consisting of the seabed, the subsoil and 
  the water column above the seabed -- that is bounded by a series of rhumb 
  lines drawn from a point 53\eqn{^\circ}{deg}03'07.6"N, 135\eqn{^\circ}{deg}50'25.9"W, to a 
  point 53\eqn{^\circ}{deg}16'20.9"N, 134\eqn{^\circ}{deg}59'55.4"W, then to a point 
  53\eqn{^\circ}{deg}39'49.2"N, 135\eqn{^\circ}{deg}17'04.9"W, then to a point 
  53\eqn{^\circ}{deg}39'18.0"N, 135\eqn{^\circ}{deg}53'46.5"W, then to a point 
  53\eqn{^\circ}{deg}52'16.7"N, 136\eqn{^\circ}{deg}30'23.1"W, then to a point 
  53\eqn{^\circ}{deg}49'19.6"N, 136\eqn{^\circ}{deg}47'33.1"W, then to a point 
  53\eqn{^\circ}{deg}40'02.5"N, 136\eqn{^\circ}{deg}57'03.5"W, then to a point 
  53\eqn{^\circ}{deg}13'59.2"N, 136\eqn{^\circ}{deg}10'00.0"W, then back to the point of 
  commencement (which points are shown as points 1 to 8, respectively, of the 
  schedule) is designated as the Bowie Seamount Marine Protected Area.
}
\source{
  Justice Laws Website \cr
  Bowie Seamount Marine Protected Area Regulations (SOR/2008-124) \cr
  \url{http://laws-lois.justice.gc.ca/eng/regulations/SOR-2008-124/page-1.html}
}
\keyword{datasets}
