\name{pl230}
\docType{data}
\alias{pl230}
\title{Topo: 230 Degree True Line from Lookout Island}
\description{
  PolySet of one polyline describing the boundary between the Longspine
  Thornyhead \emph{Sebastolobus altivelis} management regions in fishing 
  years 2000 and 2001 (April 1 to March 31).
}
\usage{data(pl230)}
\format{
  Data frame comprising 4 columns: \cr
    \code{PID...}{primary polygon ID} \cr
    \code{POS...}{position of each vertex within a given polygon} \cr
    \code{X.....}{longitude coordinates} \cr
    \code{Y.....}{latitude coordinates}
}
\source{
  Polyline data in use at Fisheries and Oceans Canada (Pacific region)
  for describing the line 230\eqn{^\circ}{deg} True from Lookout
  Island (127\eqn{^\circ}{deg} 26' 57.3" W, 49\eqn{^\circ}{deg} 59' 52.2" N) 
  that formed the boundary between the traditional
  west coast Vancouver Island Longspine Thornyhead fishing ground and
  the northern exploratory fishing grounds during the 2000 and 2001
  fishing years (Schnute \emph{et al.} 2004).
}
\references{
  Schnute, J., Haigh, R., Krishka, B., Sinclair, A. and Starr, P. (2004) 
  The British Columbia Longspine Thornyhead fishery: analysis of survey and 
  commercial data (1996-2003). \emph{Canadian Science Advisory Secretariat, 
  Research Document} \bold{2004/059}, 75 pp.\cr
  \url{http://www.dfo-mpo.gc.ca/csas/CSAS/publications/ResDocs-DocRech/2004/2004_059_E.htm}
}
\seealso{
  \code{\link{ltea}}, \code{\link{ltxa}}
}
\keyword{datasets}
