\name{spongeCPZ}
\alias{spongeCPZ}
\alias{spongeAMZ}
\docType{data}
\title{
  Topo: Sponge Reef Zones
}
\description{
  Sponge reef core protected zones and adaptive management zones within the 
  proposed MPA \emph{Hecate Strait and Queen Charlotte Sound Glass Sponge Reefs}.
}
\usage{
data(spongeCPZ)
data(spongeAMZ)
}
\format{
  \code{spongeCPZ}: a data frame with 43 observations on 5 variables (below). \cr
  \code{spongeAMZ}: a data frame with 34 observations on 4 variables (no \code{SID}).
  \describe{
    \item{\code{PID}}{primary polygon ID}
    \item{\code{SID}}{secondary polygon ID}
    \item{\code{POS}}{vertex position}
    \item{\code{X}}{geographic longitude (\eqn{^\circ}N)}
    \item{\code{Y}}{geographic latitude (\eqn{^\circ}W)}
  }
}
\details{
The Hecate Strait and Queen Charlotte Sound Glass Sponge Reefs have been 
identified as an ecologically and biologically significant area due to their 
global geological uniqueness (Conway et al. 1991, Conway et al. 2001, and 
Kruatter et al. 2001), and there is international and national recognition that 
cold-water corals and sponge dominated communities can serve as key structural 
habitat for many fish and invertebrate species (DFO 2010). 
This area is currently in the process of being designated as an Oceans Act 
Marine Protected Area (MPA) as part of the Health of the Oceans Initiative. 
It has been identified as an Area of Interest in consideration of an ecosystem-
based management (EBM) approach for the Pacific North Coast Integrated 
Management Area (PNCIMA), within which the reefs are located.

The proposed MPA consists of three separate areas totalling 2410 square km that 
include the four glass sponge reef complexes located in Hecate Strait and Queen 
Charlotte Sound, the water column and the surrounding waters, and the seabed and
subsoil to a depth of 20 meters. The three areas are referred to as the Northern
Reef, the Central Reef (containing two reef complexes), and the Southern Reef. 
Each of the three areas is proposed to have three internal management zones, 
referred to as the Core Protection Zone (CPZ), the Adaptive Management Zone 
(AMZ) and the Vertical Adaptive Management Zone (VAMZ).
}
\source{
  Lisa Lacko, Pacific Biological Station, Fisheries and Ocean Canada, Nanaimo BC
}
\references{
DFO (2012)
Identification and evaluation of biological effects and impacts of sediment to 
sponge communities in Hecate Strait Pacific regional science advisory process.
October 23-25, 2012, Nanaimo BC. \cr
\url{http://www.dfo-mpo.gc.ca/csas-sccs/Schedule-Horraire/2012/10_23-25b-eng.html}

Conway, K.W., Krautter, M., Barrie, J.V. and Neuweiler, M. (2001) 
Hexactinellid sponge reefs on the Canadian continental shelf: A unique \sQuote{living fossil}. 
\emph{Geoscience Canada} \bold{28(20)}: 71--78.

Conway, K.W., Barrie, J.V., Austin, W.C., and Luternauer, J.L. (1991) 
Holocene sponge bioherms on the western Canadian continental shelf. 
\emph{Continental Shelf Research} \bold{11(8-10)}: 771--790.

DFO. (2010) 
Pacific Region Cold-Water Coral and Sponge Conservation Strategy 2010-2015. 
Fisheries and Oceans Canada. \bold{2010/1663}.

Krautter, M., Conway, K., Barrie, J.V., and Neuweiler, M. (2001) 
Discovery of a \sQuote{living dinosaur}: globally unique modern Hexactinellid sponge reefs off British Columbia, Canada. 
\emph{Facies} \bold{44}: 265--282.
}
\keyword{datasets}

