\name{tidemarks}
\alias{tidemarks}
\docType{data}
\title{
  Topo: Tide Marks Closed Area
}
\description{
  The Tide Marks closed area serves to reduce harvesting pressure on
  Pacific Ocean Perch stocks during the spawning period.
}
\usage{data("tidemarks")}
\format{
  A data frame with 28 observations on the following 4 variables.
  \describe{
    \item{\code{PID}}{numeric primary polygon identifier}
    \item{\code{POS}}{numeric vertex counter}
    \item{\code{X}}{numeric longitude (deg.W)}
    \item{\code{Y}}{numeric latitude (deg.N)}
  }
}
\details{
  Closed to all trawling (includes both bottom and midwater) from 
  February 21 to May 31 and from October 1 to March 31 in those 
  portions of Areas 109 to 111 and Subareas 130-2, 108-2 and 130-1
  found within a line that begins at the intersection of the outer
  perimeter of Fishing Zone 5 and 51 deg 39.33 min N. lat. then
  following the northern boundary of Subarea 130-2 to 51 deg 39.33
  min N. lat. 131 deg 00 min W. long. then to 51 deg 36.00 min N. lat.
  130 deg 42.02 min W. long. then to 51 deg 48 min N. lat. 130 deg
  00 min W. long. then to 51 deg 47 min N. lat. 129 deg 37 min W.
  long. then to 51 deg 28 min N. lat. 129 deg 48 min W. long. then to
  51 deg 13 min N. lat. 129 deg 28 min W. long. then true south to 
  51 deg 04 min N. lat. 129 deg 28 min W. long. then to 50 deg 
  52 min N. lat. 129 deg 36 min W. then southern boundary of 130-1 to
  the outer perimeter of Fishing Zone 5 and back to the point of
  commencement.
}
\source{
  \url{http://www.pac.dfo-mpo.gc.ca/fm-gp/ifmp-eng.html}
}
\references{
  DFO. (2013) Pacific region Integrated Fisheries Management Plan
  (IFMP) for Groundfish, effective February 21, 2013 to April 11,
  2013 (Version 1.1), Appendix 8: 2013/2014 Groundfish Trawl 
  Commercial Harvest Plan.
}
\keyword{datasets}

