\name{trawlfoot}
\alias{trawlfoot}
\docType{data}
\title{
  Topo: BC Trawl Footprint
}
\description{
  Areas along the BC continental shelf/slope that have been vetted by DFO and 
  conservation groups for trawling. The footprint was frozen in 2012 and first 
  appears in the 2013 Integrated Fisheries Management Plan (DFO 2013).
}
\usage{data(trawlfoot)}
\format{
  A data frame with 572 observations on the following 5 variables.
  \describe{
    \item{\code{PID}}{primary polygon ID}
    \item{\code{SID}}{secondary polygon ID}
    \item{\code{POS}}{vertex position}
    \item{\code{X}}{geographic longitude (\eqn{^\circ}{deg}W)}
    \item{\code{Y}}{geographic latitude (\eqn{^\circ}{deg}N)}
  }
}
\details{
\bold{Environmentalists, trawl industry agree on conservation measures for B.C. 
corals and sponges}

\emph{By Larry Pynn, Vancouver Sun March 28, 2012}

Environmentalists and the groundfish bottom trawl industry on Wednesday 
announced a landmark agreement on ways to limit the impact of trawl nets on 
sensitive corals, sponges, and deepsea habitats on the B.C. coast.

The agreement sets out total annual bycatch objectives for the entire fleet at 
562 kg for corals and 322 kg for sponges, the lowest levels reported in the 
past 15 years.

A protocol is in place to alert skippers in the fleet when a trawl bycatch 
exceeds 20 kg of corals or sponges.

The total area of coastline trawled will also be reduced by 20.6\% 
to 31,633 km\eqn{^2}, the amount of continental slope habitat trawled to a depth
of 200 to 800 m by 18\% to 12,413 km\eqn{^2}, and the amount of deepwater 
habitat trawled to 800 to 1,400 m by 65\% to 1,395 km\eqn{^2}.

The agreement, three years in the making, also provides for a joint habitat 
conservation review committee consisting of representatives from industry, 
environmental groups and the federal fisheries department.

The David Suzuki Foundation and Living Oceans Society reached the agreement with
the Canadian Groundfish Research and Conservation Society, a commercial fishing 
industry organization representing participants in the commercial groundfish 
trawl fishery, including licence holders, vessel owners, crew, processing 
plants, and shore workers.

Federal fisheries staff provided data analysis and ensured the agreement met 
federal requirements.

Scott Wallace, sustainable fisheries analyst for the David Suzuki Foundation, 
said in an interview that there are 146 total licences in the fleet, of which 
50 vessels are most active.

\dQuote{The whole idea is to provide strong accountability at the tow-by-tow 
level, which is quite unique in fisheries,} he said.

\dQuote{Each time they put down the net, they'll be thinking, 'Will I hit 
corals or sponges here?' If they do, in any large number, they'll be in trouble.}

\dQuote{I can't think of another fishery in the world that has this level of 
personal accountability.}

Bruce Turris, executive manager of the Canadian Groundfish Research and 
Conservation Society, said his industry has taken measures to evolve over the 
past 15 years, including adopting an individual quota system and 100\% observer 
coverage of catches.

With markets demanding sustainable seafood catches, the next logical step was 
to create a management plan to minimize impact on fish habitat, he said. 
\dQuote{It was important for us to be proactive in development of management 
measures that meet market needs.}

Turris noted that the new initiative is meant to pose a disincentive to the 
capture of corals and sponges, and that there is no research to determine what 
a sustainable bycatch might be.

He hopes the effort leads to trawl-caught rockfish being removed from the 
\dQuote{Don't buy} red list as determined by the Monterey Bay Aquarium's 
Seafood Watch program, which is used as a reference by other environmental 
organizations.

The new management measures are being implemented through the federal 
groundfish integrated fisheries management plan.
}
\source{
  Norm Olsen, Pacific Biological Station, Fisheries and Ocean Canada, Nanaimo BC
}
\references{
Pynn, L. (2013) 
Environmentalists, trawl industry agree on conservation measures for B.C. corals and sponges.
Vancouver Sun, March 28, 2012. 
lpynn@vancouversun.com \cr
\url{http://www.vancouversun.com/technology/Environmentalists+bottom+trawl+industry+agree+conservation+measures+corals+sponges/6373561/story.html}

DFO. 2013. Pacific region integrated fisheries management plan: groundfish.
  Effective February 21, 2013.

}
\keyword{datasets}

